+++ crates/zng-ext-font/src/lib.rs
@@ -327,7 +327,7 @@ impl FontNames {
 
         if lang!("zh-Hans").matches(lang, true, false) {
             [
-                "-apple-system",
+                "SF Pro",
                 "PingFang SC",
                 "Hiragino Sans GB",
                 "Apple Color Emoji",
@@ -335,12 +335,12 @@ impl FontNames {
             ]
             .into()
         } else if lang!("zh-Hant").matches(lang, true, false) {
-            ["-apple-system", "PingFang TC", "Apple Color Emoji", "sans-serif"].into()
+            ["SF Pro", "PingFang TC", "Apple Color Emoji", "sans-serif"].into()
         } else if lang!(ja).matches(lang, true, false) {
-            ["-apple-system", "Hiragino Kaku Gothic Pro", "Apple Color Emoji", "sans-serif"].into()
+            ["SF Pro", "Hiragino Kaku Gothic Pro", "Apple Color Emoji", "sans-serif"].into()
         } else if lang!(ko).matches(lang, true, false) {
             [
-                "-apple-system",
+                "SF Pro",
                 "Nanum Gothic",
                 "Apple SD Gothic Neo",
                 "AppleGothic",
@@ -349,7 +349,7 @@ impl FontNames {
             ]
             .into()
         } else {
-            ["-apple-system", "Apple Color Emoji", "sans-serif"].into()
+            ["SF Pro", "Apple Color Emoji", "sans-serif"].into()
         }
     }